requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
mysqlclient==2.0.3
pytz==2020.5
sqlparse==0.4.1
